Fix swapped pv1<->pv2 modes (change by gowima)
Byte order in the implementation of ByteH|L and EnumH|L
is big endian, which is in accordance to the GoodWe Modbus
register documentation. So the byte sequence left to right
in a register is H - L. Modes are stored as a byte sequence
pv4_mode ... pv1_mode. ---> ByteH/ByteL have to be exchanged
pairwase for mode 4, 3 and 2, 1.